Pull from original repository

+++ pyNastran/op2/tables/oes_stressStrain/complex/oes_bush1d.py
@@ -44,7 +44,7 @@ class ComplexCBush1DArray(OES_Object):
         #print("ntimes=%s nelements=%s ntotal=%s" % (self.ntimes, self.nelements, self.ntotal))
         dtype, idtype, fdtype = get_times_dtype(self.nonlinear_factor, self.size, self.analysis_fmt)
         self._times = np.zeros(self.ntimes, dtype=self.analysis_fmt)
-        self.element = np.zeros(self.nelements, dtype='int32')
+        self.element = np.zeros(self.nelements, dtype=idtype)
 
         #[tx, ty, tz, rx, ry, rz]
         self.data = np.zeros((self.ntimes, self.nelements, 6), dtype='complex64')
